+++ demo/pages/sdk/actions/getCallQuote.tsx
@@ -1,9 +1,9 @@
 import { NextPage } from 'next'
 import { useState } from 'react'
-import { Execute } from '@reservoir0x/relay-sdk'
+import { Execute } from '@relayprotocol/relay-sdk'
 import { useWalletClient } from 'wagmi'
 import { base, baseGoerli, sepolia, zora } from 'viem/chains'
-import { useRelayClient } from '@reservoir0x/relay-kit-ui'
+import { useRelayClient } from '@relayprotocol/relay-kit-ui'
 import { zeroAddress } from 'viem'
 import { ConnectButton } from 'components/ConnectButton'