+++ packages/ui/src/components/primitives/AccessibleList.tsx
@@ -4,7 +4,7 @@ import {
   cva,
   css as designCss,
   type Styles
-} from '@reservoir0x/relay-design-system/css'
+} from '@relayprotocol/relay-design-system/css'
 
 const AccessibleListContainerCss = cva({
   base: {