+++ packages/ui/src/types/RelayTransaction.ts
@@ -1,4 +1,4 @@
-import type { paths } from '@reservoir0x/relay-sdk'
+import type { paths } from '@relayprotocol/relay-sdk'
 
 export type UserTransactionsResponse =
   paths['/requests/v2']['get']['responses']['200']['content']['application/json']