+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @reservoir0x/relay-kit-ui
 
+## 2.17.3
+
+### Patch Changes
+
+- b2d6c8c: Always show price impact regardless of input mode
+
+## 2.17.2
+
+### Patch Changes
+
+- 8a8d560: Reenable solana for protocolv2
+
+## 2.17.1
+
+### Patch Changes
+
+- a50bea6: Disable protocol v2 on solana
+
+## 2.17.0
+
+### Minor Changes
+
+- 6400542: Add websocket support and refactor executeSteps
+
+### Patch Changes
+
+- Updated dependencies [6400542]
+  - @reservoir0x/relay-kit-hooks@1.12.0
+  - @reservoir0x/relay-sdk@2.4.0
+
 ## 2.16.2
 
 ### Patch Changes